# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for sending notifications to users",
        "title": "Notifier API",
        "contact": {
            "name": "Dima Kalinin",
            "url": "https://dimdimi4-and-fixiki.github.io/dima-portal/",
            "email": "kalinindima123@yandex.ru"
        },
        "version": "0.0.1"
    },
    "basePath": "/api/",
    "paths": {
        "/admin/get_user/": {
            "post": {
                "description": "Return user by user id or user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "get user info",
                "parameters": [
                    {
                        "description": "no_comm",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.GetUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RespErr"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.RespErr"
                        }
                    }
                }
            }
        },
        "/admin/register_user/": {
            "post": {
                "description": "Register new user with user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "User details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Resp-entity_UserCommonInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.RespErr"
                        }
                    }
                }
            }
        },
        "/v1/send_proposal/": {
            "post": {
                "description": "Sends message via email to owner of project and confirmation to user's email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1"
                ],
                "summary": "send proposal notification by email",
                "parameters": [
                    {
                        "description": "no_comm",
                        "name": "proposal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SendProposalReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RespErr"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.RespErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.GetUserReq": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                },
                "user_login": {
                    "type": "string"
                }
            }
        },
        "entity.SendProposalReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.SendProposalResp": {
            "type": "object",
            "properties": {
                "proposal_description": {
                    "type": "string"
                },
                "proposal_id": {
                    "type": "string"
                },
                "proposal_reason": {
                    "type": "string"
                },
                "proposal_title": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "cred": {
                    "$ref": "#/definitions/entity.UserCred"
                },
                "info": {
                    "$ref": "#/definitions/entity.UserCommonInfo"
                }
            }
        },
        "entity.UserCommonInfo": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/entity.UserRole"
                }
            }
        },
        "entity.UserCred": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "entity.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "service",
                "person"
            ],
            "x-enum-varnames": [
                "Admin",
                "Service",
                "Person"
            ]
        },
        "http.ErrForResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "http.Resp-entity_SendProposalResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "meta": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/entity.SendProposalResp"
                }
            }
        },
        "http.Resp-entity_UserCommonInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "meta": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/entity.UserCommonInfo"
                }
            }
        },
        "http.RespErr": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/http.ErrForResp"
                },
                "meta": {
                    "type": "string"
                }
            }
        }
    }
}